streaming/requirements.txt
nameko
nameko-grpc
python-dotenv
pytest
pytest-cov
grpcio-tools